Renamed math.h functions 'funcf.*' is float, 'funcl.*' is long double, 'func.*' is float and long double
