+++ src/libc/acoshl.c
@@ -1,0 +1,6 @@
+#include <math.h>
+
+long double acoshl(long double x)
+{
+    return logl(x + sqrtl(x * x - 1));
+}
