+++ src/libc/asinf.c
