+++ src/libc/asinhl.c
@@ -1,0 +1,6 @@
+#include <math.h>
+
+long double asinhl(long double x)
+{
+    return logl(x + sqrtl(x * x + 1));
+}
