+++ src/libc/atanhf.c
@@ -6,8 +6,3 @@ float atanhf(float x)
 }
 
 double atanh(double) __attribute__((alias("atanhf")));
-
-long double atanhl(long double x)
-{
-    return .5l * logl((1 + x) / (1 - x));
-}
