+++ src/libc/atanhl.c
@@ -1,0 +1,6 @@
+#include <math.h>
+
+long double atanhl(long double x)
+{
+    return .5l * logl((1 + x) / (1 - x));
+}
