+++ src/libc/cosf.src
