+++ src/libc/coshf.c
