+++ src/libc/coshf.src
