+++ src/libc/erff.c
@@ -6,8 +6,3 @@ float erff(float x)
 }
 
 double erf(double) __attribute__((alias("erff")));
-
-long double erfl(long double x)
-{
-    return 1 - erfcl(x);
-}
