+++ src/libc/erfl.c
@@ -1,0 +1,6 @@
+#include <math.h>
+
+long double erfl(long double x)
+{
+    return 1 - erfcl(x);
+}
