+++ src/libc/exp2f.c
