+++ src/libc/expm1f.c
