+++ src/libc/lgammaf.c
