+++ src/libc/log10f.c
