+++ src/libc/log1pf.c
