+++ src/libc/log2f.c
