+++ src/libc/logf.c
