+++ src/libc/powf.c
