+++ src/libc/remquof.c
