+++ src/libc/sinf.c
