+++ src/libc/sinf.src
