+++ src/libc/sinhf.c
