+++ src/libc/sinhf.src
