+++ src/libc/tanhf.c
