+++ src/libc/tanhf.src
